+++ requests_mv_integrations/__init__.py
@@ -21,9 +21,6 @@ from requests_mv_integrations.support.tune_request import (
 from .request_mv_integration import (
     RequestMvIntegration,
 )
-from .request_mv_integration_download import (
-    RequestMvIntegrationDownload
-)
 from .request_mv_integration_upload import (
     RequestMvIntegrationUpload
 )
